chore: moved test_command_processor.py into processor tests folder.
